Move stuff out of valuenum.h
