+++ src/coreclr/jit/jitconfigvalues.h
@@ -390,7 +390,7 @@ CONFIG_INTEGER(JitOptRepeatCount, W("JitOptRepeatCount"), 2) // Number of times 
 #endif                                                       // defined(OPT_CONFIG)
 
 // Max # of MapSelect's considered for a particular top-level invocation.
-CONFIG_INTEGER(JitVNMapSelBudget, W("JitVNMapSelBudget"), DEFAULT_MAP_SELECT_BUDGET)
+CONFIG_INTEGER(JitVNMapSelBudget, W("JitVNMapSelBudget"), DefaultVNMapSelectBudget)
 
 CONFIG_INTEGER(TailCallLoopOpt, W("TailCallLoopOpt"), 1) // Convert recursive tail calls to loops
 CONFIG_METHODSET(AltJit, W("AltJit"))         // Enables AltJit and selectively limits it to the specified methods.
